+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Scripts to do releases
+- pyproject.toml file
+
 ## [0.7.0] - 2026-02-10
 
 ### Added