+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2026-03-17
+
 ### Added
 
 - Scripts to do releases
 - pyproject.toml file
-- Add EPA Dscenario endpoints
+- EPA Dscenario endpoints
 
 ### Changed
 
@@ -160,7 +162,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic test with pytest.
 - Basic CI workflow.
 
-[unreleased]: https://github.com/giswater/giswater-api/compare/v0.7.0...main
+[unreleased]: https://github.com/giswater/giswater-api/compare/v0.8.0...main
+[0.8.0]: https://github.com/giswater/giswater-api/compare/v0.7.0...0.8.0
 [0.7.0]: https://github.com/giswater/giswater-api/compare/v0.6.0...0.7.0
 [0.6.0]: https://github.com/giswater/giswater-api/compare/v0.5.0...0.6.0
 [0.5.0]: https://github.com/giswater/giswater-api/compare/v0.4.0...v0.5.0